Add all models to match the supplement material
JB sent me all the models she used for the results reported in the supplement material and I have updated the repository. I also decided to upload the models as models.RData so that they can be easily inspected for summary statistics. There were also some updates in readme files.